32/64 build of OAuthConsumer.framework
Tracking changes here: http://bitbucket.org/sholt/oauth/

+++ Frameworks/OAuthConsumer.framework/Versions/A/Headers/OAToken.h
@@ -36,6 +36,6 @@
 - (id)initWithKey:(NSString *)aKey secret:(NSString *)aSecret;
 - (id)initWithUserDefaultsUsingServiceProviderName:(NSString *)provider prefix:(NSString *)prefix;
 - (id)initWithHTTPResponseBody:(NSString *)body;
-- (int)storeInUserDefaultsWithServiceProviderName:(NSString *)provider prefix:(NSString *)prefix;
+- (NSInteger)storeInUserDefaultsWithServiceProviderName:(NSString *)provider prefix:(NSString *)prefix;
 
 @end
